Refactoring: add ListPosts and WritePost component. Improvements: add Flow.
