+++ webapp/src/writePost.js
@@ -1,0 +1,18 @@
+import React from "react";
+import logo from "./logo.svg";
+import "./writePost.css";
+
+function WritePost() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Testing <code>writePost.js</code>.
+        </p>
+      </header>
+    </div>
+  );
+}
+
+export default WritePost;
